+++ tests/integration/cases/call_ref_nested_converted/Elixir_call.ex
@@ -1,0 +1,7 @@
+defmodule Check do
+  def process(_data), do: nil
+  def x do
+    my_var = 42
+    process([my_var, 42, "static"])
+  end
+end