+++ tests/integration/cases/call_ref_nested_converted/Php_call.php
@@ -1,0 +1,4 @@
+<?php
+function process($data) {}
+$my_var = 42;
+process(data: [$my_var, 42, "static"]);