+++ tests/integration/cases/call_ref_nested_converted/Swift_call.swift
@@ -1,0 +1,3 @@
+@discardableResult func process(data: Any = 0) -> Any { 0 }
+let myVar: Any = 42
+process(data: [myVar, 42, "static"]);