+++ tests/integration/cases/call_ref_nested_converted/VisualBasic_call.vb
@@ -1,0 +1,10 @@
+Imports System.Collections.Generic
+Module Check
+    Function process(data As Object) As Object
+        Return Nothing
+    End Function
+    Sub _calls()
+        Dim MyVar = 42
+        process(New Object() {MyVar, 42, "static"})
+    End Sub
+End Module