+++ tests/integration/cases/call_ref_nested_in_dict/Cpp_call.cpp
@@ -1,0 +1,10 @@
+#include <initializer_list>
+#include <string>
+#include <map>
+#include <vector>
+auto process(auto...) { return 0; }
+int main() {
+auto my_var = 42;
+process(std::map<std::string, std::variant<std::map<std::string, std::string>, int>>{{"key", std::move(my_var)}, {"count", 42}});
+    return 0;
+}