+++ tests/integration/cases/call_ref_nested_in_dict/Elixir_call.ex
@@ -1,0 +1,7 @@
+defmodule Check do
+  def process(_data), do: nil
+  def x do
+    my_var = 42
+    process(%{"key" => my_var, "count" => 42})
+  end
+end