+++ tests/integration/cases/call_ref_nested_in_dict/PowerShell_call.ps1
@@ -1,0 +1,3 @@
+function process {}
+$my_var = 42
+process(@{"key" = $my_var; "count" = 42})