+++ tests/integration/cases/call_ref_nested_in_dict/V_call.v
@@ -1,0 +1,7 @@
+interface ICallArg_ {}
+fn process(args ...ICallArg_) {}
+
+fn main() {
+	my_var := 42
+	process({'key': IVal(my_var.clone()), 'count': IVal(42)});
+}